Add ability to auto share and accept quests
This adds new event handlers for:

QUEST_ACCEPT_CONFIRM
QUEST_DETAIL
QUEST_ACCEPTED

Which will either automatically accept quests, or
when one accepting a new quest will attempt to
share it with others in the party. The events
trigger a cycle, however the game client will
refuse to share if the other character is already
on the quest. End result is pretty clean and
simplifies questing with >1 character at a
time.

This does not help with quest turn ins. As-is
each character needs to interact and turn the
finished quests in separately.